added generic mongo repository, using bigDecimal as ids, added methods to repositories

+++ src/main/java/ro/danix/first/controller/converter/UserConverter.java
@@ -1,5 +1,6 @@
 package ro.danix.first.controller.converter;
 
+import java.math.BigDecimal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.converter.Converter;
 import ro.danix.first.model.domain.user.User;
@@ -16,6 +17,6 @@ public class UserConverter implements Converter<String, User> {
 
     @Override
     public User convert(String source) {
-        return this.userService.findOne(new Long(source));
+        return this.userService.findOne(new BigDecimal(source));
     }
 }
