+++ src/main/java/ro/danix/first/model/repository/user/UserRepository.java
@@ -1,5 +1,6 @@
 package ro.danix.first.model.repository.user;
 
+import java.math.BigDecimal;
 import org.springframework.data.repository.Repository;
 import ro.danix.first.model.domain.EmailAddress;
 import ro.danix.first.model.domain.user.User;
@@ -8,9 +9,9 @@ import ro.danix.first.model.domain.user.User;
  *
  * @author danix
  */
-public interface UserRepository extends Repository<User, Long> {
+public interface UserRepository extends Repository<User, BigDecimal> {
 
-    User findOne(Long id);
+    User findOne(BigDecimal id);
 
     User save(User user);
